Efficiency fixes
ASDver is now a dictionary with a key of the version name and the
values being a list of strings. Stopped using classes for each ASD
version, as I had to use eval() to loop through and search each class,
which can pose a security risk if passed the wrong arguments.